[TT-10781] Split openapi.go file
